requirements.txt
# Core
fastapi>=0.115
uvicorn>=0.37
openai>=1.0.0
pydantic>=2.7,<3
psycopg2-binary>=2.9,<3
requests>=2.26.0
python-dotenv>=1.0
click>=8.0

# Performance / server extras (optional)
httptools>=0.6
uvloop>=0.21
watchfiles>=1.1
websockets>=15.0

# Text / tokenization / validation
tiktoken>=0.5
anyio>=4.0
email-validator>=2.0